+++ cmd/doc.go
@@ -335,33 +335,6 @@ func diagExists(cfg *config.Config) error {
 	return nil
 }
 
-func newConfig() (*config.Config, error) {
-	cfg := config.New()
-	if err := cfg.LoadConfigFile(detectConfigPath(configPath)); err != nil {
-		return nil, err
-	}
-	if len(nodeLists) == 0 {
-		cfg.HideRealNodes = true
-	} else {
-		for _, n := range nodeLists {
-			if err := cfg.LoadRealNodesFile(n); err != nil {
-				return nil, err
-			}
-		}
-	}
-	if err := cfg.Build(); err != nil {
-		return nil, err
-	}
-
-	if hideDetails {
-		if err := cfg.HideDetails(); err != nil {
-			return nil, err
-		}
-	}
-
-	return cfg, nil
-}
-
 func init() {
 	docCmd.Flags().BoolVarP(&force, "force", "", false, "generate a document without checking for the existence of an existing document")
 	docCmd.Flags().StringVarP(&configPath, "config", "c", "", "config file path")
